[ALL] Converted variables with expected bit-age to uint32/64_t and uintptr_t.
[GAFF]
Made the FNV "wrapper" functions just plain old overloads and removed the default argument for the versions that take an init value.
Added custom assert handler API.

[SHIBBOLETH]
Fixed rawRequestInterface() interfaces to take a Gaff::ReflectionHash instead of an unsigned int.

TODO: Convert assert() calls to GAFF_ASSERT()/GAFF_ASSERT_MSG() calls.

+++ dependencies/Gaff/include/Gaff_MetaData.h
@@ -57,7 +57,7 @@ public:
 
 	INLINE bool getBool(unsigned int index) const;
 	INLINE int getInt(unsigned int index) const;
-	INLINE unsigned long long getUInt64(unsigned int index) const;
+	INLINE uint64_t getUInt64(unsigned int index) const;
 	INLINE float getFloat(unsigned int index) const;
 	INLINE const char* getString(unsigned int index) const;
 	INLINE const float* getVec3(unsigned int index) const;
