+++ src/Game/Shibboleth_ModelLoader.cpp
@@ -74,7 +74,7 @@ ModelLoader::~ModelLoader(void)
 {
 }
 
-Gaff::IVirtualDestructor* ModelLoader::load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map)
+Gaff::IVirtualDestructor* ModelLoader::load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map)
 {
 	assert(file_map.hasElementWithKey(AString(file_name)));
