+++ src/Game/Shibboleth_TextureLoader.cpp
@@ -43,7 +43,7 @@ TextureLoader::~TextureLoader(void)
 {
 }
 
-Gaff::IVirtualDestructor* TextureLoader::load(const char* file_name, unsigned long long, HashMap<AString, IFile*>& file_map)
+Gaff::IVirtualDestructor* TextureLoader::load(const char* file_name, uint64_t, HashMap<AString, IFile*>& file_map)
 {
 	assert(file_map.hasElementWithKey(AString(file_name)));
